Moved more functions to common
