+++ src/common.h
@@ -2,5 +2,7 @@ void printVersion(char* programName);
 void printMenu(int line, int col, char *menustring);
 void topLineMessage(const char *message);
 int readline(char *buffer, int buflen, char *oldbuf);
+int check_dir(char *pwd);
+int check_file(char *file);
 void clear_workspace();
 int exittoshell();
